Deal with unknown errors.
git-svn-id: http://source.meta.net.nz/svn/mp3parser/trunk@46 4640f49e-d398-488d-9bdf-65a890809393

+++ flacnamer.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python
+#!/usr/bin/python2.5
 #
 # Script to automatically look up albums in the musicbrainz database and
 # rename/retag FLAC files.
@@ -34,6 +34,7 @@ import musicdns
 
 AMAZON_LICENSE_KEY='1WQQTEA14HEA9AERDMG2'
 MUSICDNS_KEY='a7f6063296c0f1c9b75c7f511861b89b'
+lastwsquery = time.time()
 
 def print_usage():
 	print "usage: " + sys.argv[0] + " <srcpath> [OPTIONS]"
@@ -85,19 +86,24 @@ def get_releases_by_metadata(disc):
 	# Filter out of the list releases with a different number of tracks to the
 	# Disc.
 	for rel in rels:
-		release = get_release_by_releaseid(rel.id)
+		release =rel.release  #get_release_by_releaseid(rel.id)
 		if len(release.getTracks()) == len(disc.tracks):
 			releases.append(rel)
 
 	return releases
 
+
 def get_tracks_by_puid(puid):
 	""" Lookup a list of musicbrainz tracks by PUID. Returns a list of Track
 	objects. """ 
+	global lastwsquery
 	q = ws.Query()
 	filter = ws.TrackFilter(puid=puid)
 	results = []
+	if time.time()-lastwsquery<2:
+		time.sleep(time.time()-lastwsquery)
 	rs = q.getTracks(filter=filter)
+	lastwsquery=time.time()
 	for r in rs:
 		results.append(r.getTrack())
 	return results
@@ -127,11 +133,10 @@ def get_release_by_fingerprints(disc):
 	for t in disc.tracks:
 		tracknum += 1
 
-		print "Decoding..."
 		tmp = os.tmpnam() + ".wav"
-		os.system("flac -d --totally-silent -o " +  tmp +  " " + t.filename)
+		if os.system("flac -d --totally-silent -o " +  tmp +  " " + t.filename)!=0:
+			raise Exception("flac %s failed!" % t.filanem )
 
-		print "Fingerprinting..."
 		(fp, duration) = fingerprint.fingerprint(tmp)
 		(artist, trackname, puid) = musicdns.lookup_fingerprint(fp, duration, MUSICDNS_KEY)
 		os.unlink(tmp)
@@ -143,10 +148,10 @@ def get_release_by_fingerprints(disc):
 
 		tracks = get_tracks_by_puid(puid)
 		for track in tracks:
-			print "  Could be " + track.id
+			print "  Could be " + track.id + " (%s)" % track.title
 			releases = track.getReleases()
 			for r in releases:
-				print "     Which is on " + r.id
+				print "     Which is on " + r.id + " (%s)" % r.title
 				release = get_release_by_releaseid(r.id)
 
 				# Filter releases with the wrong number of tracks
@@ -328,6 +333,8 @@ def main():
 
 	for i in range(len(disc.tracks)):
 		disc.tracks[i].filename = os.path.join(srcpath,  "track" + str(i + 1).zfill(2) + ".flac")
+		if not os.path.exists(disc.tracks[i].filename):
+			disc.tracks[i].filename = os.path.join(srcpath,  "track" + str(i + 1).zfill(2) + ".cdda.flac")
 
 	print "discID: " + disc.discid
 
@@ -395,7 +402,6 @@ def main():
 
 	if (os.path.exists(newpath)):
 		print "Destination path already exists, skipping" 
-		time.sleep(1)
 		sys.exit(3)
 
 	os.mkdir(newpath)
@@ -507,5 +513,4 @@ DISCTOTAL=%s
 if __name__ == "__main__":
 	main()
 	print "Success"
-	time.sleep(1)
 	sys.exit(0)
